divided processes in modules
